Add the skeleton of a CE-Go buildpack.
This pulls in a one-off build of the go-build buildpack with https://github.com/paketo-buildpacks/go-build/pull/92 that enables us to have our function step direct the target that go-build builds.

With this users can define a function named "Receiver" in the root of the build context that adheres to one of the compatible sdk-go receiver signatures.

Alternately, the user can specify a function in a sub-package using `CE_GO_PACKAGE`, and customize the function name with `CE_GO_FUNCTION`.